@@ -1,6 +1,4 @@
-﻿using System;
-using System.Configuration;
-using Microsoft.Azure.Documents.Client;
+﻿using Microsoft.Azure.Documents.Client;
 
 namespace NCS.DSS.Goal.Cosmos.Client
 {
