@@ -1,7 +1,4 @@
-﻿using System;
-using System.Threading.Tasks;
-
-namespace NCS.DSS.Goal.Cosmos.Helper
+﻿namespace NCS.DSS.Goal.Cosmos.Helper
 {
     public interface IResourceHelper
     {
