@@ -1,11 +1,10 @@
-﻿using System;
-using DFC.JSON.Standard;
+﻿using DFC.JSON.Standard;
 using NCS.DSS.Goal.Models;
 using NCS.DSS.Goal.PatchGoalHttpTrigger.Service;
 using NCS.DSS.Goal.ReferenceData;
 using Newtonsoft.Json;
 using NUnit.Framework;
-using System.Threading.Tasks;
+using System;
 
 namespace NCS.DSS.Goal.Tests.ServicesTests
 {
@@ -34,86 +33,86 @@ namespace NCS.DSS.Goal.Tests.ServicesTests
             var result = _goalPatchService.Patch(string.Empty, _goalPatch);
 
             // Assert
-            Assert.IsNull(result);
+            Assert.That(result, Is.Null);
         }
 
         [Test]
         public void GoalPatchServiceTests_CheckDateGoalAchievedIsUpdated_WhenPatchIsCalled()
         {
-            var goalPatch = new GoalPatch() {  DateGoalAchieved = DateTime.MaxValue };
+            var goalPatch = new GoalPatch() { DateGoalAchieved = DateTime.MaxValue };
 
             var patchedGoal = _goalPatchService.Patch(_json, goalPatch);
 
             var goal = JsonConvert.DeserializeObject<Models.Goal>(patchedGoal);
 
             // Assert
-            Assert.AreEqual(DateTime.MaxValue, goal.DateGoalAchieved);
+            Assert.That(goal.DateGoalAchieved, Is.EqualTo(DateTime.MaxValue));
         }
 
 
         [Test]
         public void GoalPatchServiceTests_CheckDateGoalCapturedIsUpdated_WhenPatchIsCalled()
         {
-            var goalPatch = new GoalPatch {  DateGoalCaptured = DateTime.MaxValue };
+            var goalPatch = new GoalPatch { DateGoalCaptured = DateTime.MaxValue };
 
             var patchedGoal = _goalPatchService.Patch(_json, goalPatch);
 
             var goal = JsonConvert.DeserializeObject<Models.Goal>(patchedGoal);
 
             // Assert
-            Assert.AreEqual(DateTime.MaxValue, goal.DateGoalCaptured);
+            Assert.That(goal.DateGoalCaptured, Is.EqualTo(DateTime.MaxValue));
         }
 
         [Test]
         public void GoalPatchServiceTests_CheckDateGoalShouldBeCompletedByWhenPatchIsCalled()
         {
-            var goalPatch = new GoalPatch {  DateGoalShouldBeCompletedBy = DateTime.MaxValue };
+            var goalPatch = new GoalPatch { DateGoalShouldBeCompletedBy = DateTime.MaxValue };
 
             var patchedGoal = _goalPatchService.Patch(_json, goalPatch);
 
             var goal = JsonConvert.DeserializeObject<Models.Goal>(patchedGoal);
 
             // Assert
-            Assert.AreEqual(DateTime.MaxValue, goal.DateGoalShouldBeCompletedBy);
+            Assert.That(goal.DateGoalShouldBeCompletedBy, Is.EqualTo(DateTime.MaxValue));
         }
 
         [Test]
         public void GoalPatchServiceTests_CheckDateGoalentToCustomerIsUpdated_WhenPatchIsCalled()
         {
-            var goalPatch = new Models.GoalPatch {  GoalStatus = GoalStatus.Achieved };
+            var goalPatch = new Models.GoalPatch { GoalStatus = GoalStatus.Achieved };
 
             var patchedGoal = _goalPatchService.Patch(_json, goalPatch);
 
             var goal = JsonConvert.DeserializeObject<Models.Goal>(patchedGoal);
 
             // Assert
-            Assert.AreEqual(GoalStatus.Achieved, goal.GoalStatus);
+            Assert.That(goal.GoalStatus, Is.EqualTo(GoalStatus.Achieved));
         }
 
         [Test]
         public void GoalPatchServiceTests_CheckGoalSummaryIsUpdated_WhenPatchIsCalled()
         {
-            var goalPatch = new GoalPatch {  GoalSummary  = "Summary" };
-            
+            var goalPatch = new GoalPatch { GoalSummary = "Summary" };
+
             var patchedGoal = _goalPatchService.Patch(_json, goalPatch);
 
             var goal = JsonConvert.DeserializeObject<Models.Goal>(patchedGoal);
 
             // Assert
-            Assert.AreEqual("Summary", goal.GoalSummary);
+            Assert.That(goal.GoalSummary, Is.EqualTo("Summary"));
         }
 
         [Test]
         public void GoalPatchServiceTests_CheckDateGoalAcknowledgedIsUpdated_WhenPatchIsCalled()
         {
-            var goalPatch = new Models.GoalPatch {  GoalType = GoalType.Other };
+            var goalPatch = new Models.GoalPatch { GoalType = GoalType.Other };
 
             var patchedGoal = _goalPatchService.Patch(_json, goalPatch);
 
             var goal = JsonConvert.DeserializeObject<Models.Goal>(patchedGoal);
 
             // Assert
-            Assert.AreEqual(GoalType.Other, goal.GoalType);
+            Assert.That(goal.GoalType, Is.EqualTo(GoalType.Other));
         }
 
 
@@ -121,26 +120,26 @@ namespace NCS.DSS.Goal.Tests.ServicesTests
         public void GoalPatchServiceTests_CheckLastModifiedDateIsUpdated_WhenPatchIsCalled()
         {
             var goalPatch = new GoalPatch { LastModifiedDate = DateTime.MaxValue };
-            
+
             var patchedGoal = _goalPatchService.Patch(_json, goalPatch);
 
             var goal = JsonConvert.DeserializeObject<Models.Goal>(patchedGoal);
 
             // Assert
-            Assert.AreEqual(DateTime.MaxValue, goal.LastModifiedDate);
+            Assert.That(goal.LastModifiedDate, Is.EqualTo(DateTime.MaxValue));
         }
 
         [Test]
         public void GoalPatchServiceTests_CheckLastModifiedTouchpointIdIsUpdated_WhenPatchIsCalled()
         {
             var goalPatch = new GoalPatch { LastModifiedBy = "0000000111" };
-            
+
             var patchedGoal = _goalPatchService.Patch(_json, goalPatch);
 
             var goal = JsonConvert.DeserializeObject<Models.Goal>(patchedGoal);
-   
+
             // Assert
-            Assert.AreEqual("0000000111", goal.LastModifiedBy);
+            Assert.That(goal.LastModifiedBy, Is.EqualTo("0000000111"));
         }
 
     }
