@@ -1,9 +1,9 @@
-﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations;
-using NCS.DSS.Goal.ReferenceData;
+﻿using NCS.DSS.Goal.ReferenceData;
 using NCS.DSS.Goal.Validation;
 using NUnit.Framework;
+using System;
+using System.Collections.Generic;
+using System.ComponentModel.DataAnnotations;
 
 namespace NCS.DSS.Goal.Tests.ValidationTests
 {
@@ -22,7 +22,8 @@ namespace NCS.DSS.Goal.Tests.ValidationTests
         [Test]
         public void ValidateTests_ReturnValidationResult_DateGoalCapturedMustBeLessThanDateGoalCompletedByAndCurrentDate()
         {
-            var goal = new Models.Goal {
+            var goal = new Models.Goal
+            {
                 DateGoalCaptured = DateTime.Today.AddDays(-4),
                 DateGoalShouldBeCompletedBy = DateTime.Today.AddDays(-5),
                 GoalSummary = "Summary",
@@ -33,9 +34,9 @@ namespace NCS.DSS.Goal.Tests.ValidationTests
             var result = _validate.ValidateResource(goal, false);
 
             // Assert
-            Assert.IsInstanceOf<List<ValidationResult>>(result);
-            Assert.IsNotNull(result);
-            Assert.AreEqual(1, result.Count);
+            Assert.That(result, Is.InstanceOf<List<ValidationResult>>());
+            Assert.That(result, Is.Not.Null);
+            Assert.That(result.Count, Is.EqualTo(1));
         }
 
         [Test]
@@ -54,9 +55,9 @@ namespace NCS.DSS.Goal.Tests.ValidationTests
             var result = _validate.ValidateResource(goal, false);
 
             // Assert
-            Assert.IsInstanceOf<List<ValidationResult>>(result);
-            Assert.IsNotNull(result);
-            Assert.AreEqual(1, result.Count);
+            Assert.That(result, Is.InstanceOf<List<ValidationResult>>());
+            Assert.That(result, Is.Not.Null);
+            Assert.That(result.Count, Is.EqualTo(1));
         }
 
         [Test]
@@ -74,9 +75,9 @@ namespace NCS.DSS.Goal.Tests.ValidationTests
             var result = _validate.ValidateResource(goal, false);
 
             // Assert
-            Assert.IsInstanceOf<List<ValidationResult>>(result);
-            Assert.IsNotNull(result);
-            Assert.AreEqual(1, result.Count);
+            Assert.That(result, Is.InstanceOf<List<ValidationResult>>());
+            Assert.That(result, Is.Not.Null);
+            Assert.That(result.Count, Is.EqualTo(1));
         }
 
         [Test]
@@ -88,16 +89,16 @@ namespace NCS.DSS.Goal.Tests.ValidationTests
                 DateGoalShouldBeCompletedBy = DateTime.Today.AddDays(-3),
                 GoalSummary = "Summary",
                 GoalType = GoalType.Other,
-                GoalStatus= (GoalStatus)1000,
+                GoalStatus = (GoalStatus)1000,
                 LastModifiedDate = DateTime.UtcNow
             };
 
             var result = _validate.ValidateResource(goal, false);
 
             // Assert
-            Assert.IsInstanceOf<List<ValidationResult>>(result);
-            Assert.IsNotNull(result);
-            Assert.AreEqual(1, result.Count);
+            Assert.That(result, Is.InstanceOf<List<ValidationResult>>());
+            Assert.That(result, Is.Not.Null);
+            Assert.That(result.Count, Is.EqualTo(1));
         }
 
 
@@ -116,9 +117,9 @@ namespace NCS.DSS.Goal.Tests.ValidationTests
             var result = _validate.ValidateResource(goal, false);
 
             // Assert
-            Assert.IsInstanceOf<List<ValidationResult>>(result);
-            Assert.IsNotNull(result);
-            Assert.AreEqual(1, result.Count);
+            Assert.That(result, Is.InstanceOf<List<ValidationResult>>());
+            Assert.That(result, Is.Not.Null);
+            Assert.That(result.Count, Is.EqualTo(1));
         }
 
         [Test]
@@ -135,9 +136,9 @@ namespace NCS.DSS.Goal.Tests.ValidationTests
 
             var result = _validate.ValidateResource(goal, false);
 
-            Assert.IsInstanceOf<List<ValidationResult>>(result);
-            Assert.IsNotNull(result);
-            Assert.AreEqual(0, result.Count);
+            Assert.That(result, Is.InstanceOf<List<ValidationResult>>());
+            Assert.That(result, Is.Not.Null);
+            Assert.That(result.Count, Is.EqualTo(0));
         }
 
         [Test]
@@ -154,9 +155,9 @@ namespace NCS.DSS.Goal.Tests.ValidationTests
 
             var result = _validate.ValidateResource(goal, false);
 
-            Assert.IsInstanceOf<List<ValidationResult>>(result);
-            Assert.IsNotNull(result);
-            Assert.AreEqual(1, result.Count);
+            Assert.That(result, Is.InstanceOf<List<ValidationResult>>());
+            Assert.That(result, Is.Not.Null);
+            Assert.That(result.Count, Is.EqualTo(1));
         }
 
         [Test]
@@ -173,9 +174,9 @@ namespace NCS.DSS.Goal.Tests.ValidationTests
 
             var result = _validate.ValidateResource(goal, false);
 
-            Assert.IsInstanceOf<List<ValidationResult>>(result);
-            Assert.IsNotNull(result);
-            Assert.AreEqual(0, result.Count);
+            Assert.That(result, Is.InstanceOf<List<ValidationResult>>());
+            Assert.That(result, Is.Not.Null);
+            Assert.That(result.Count, Is.EqualTo(0));
         }
 
         [Test]
@@ -192,9 +193,9 @@ namespace NCS.DSS.Goal.Tests.ValidationTests
 
             var result = _validate.ValidateResource(goal, false);
 
-            Assert.IsInstanceOf<List<ValidationResult>>(result);
-            Assert.IsNotNull(result);
-            Assert.AreEqual(1, result.Count);
+            Assert.That(result, Is.InstanceOf<List<ValidationResult>>());
+            Assert.That(result, Is.Not.Null);
+            Assert.That(result.Count, Is.EqualTo(1));
         }
 
 
@@ -207,15 +208,15 @@ namespace NCS.DSS.Goal.Tests.ValidationTests
                 DateGoalShouldBeCompletedBy = DateTime.Today.AddDays(-3),
                 GoalSummary = "Summary",
                 GoalType = GoalType.Learning,
-                LastModifiedBy = "0000000001", 
+                LastModifiedBy = "0000000001",
                 SubcontractorId = "123456"
             };
 
             var result = _validate.ValidateResource(goal, false);
 
-            Assert.IsInstanceOf<List<ValidationResult>>(result);
-            Assert.IsNotNull(result);
-            Assert.AreEqual(0, result.Count);
+            Assert.That(result, Is.InstanceOf<List<ValidationResult>>());
+            Assert.That(result, Is.Not.Null);
+            Assert.That(result.Count, Is.EqualTo(0));
         }
 
         [Test]
@@ -233,9 +234,9 @@ namespace NCS.DSS.Goal.Tests.ValidationTests
 
             var result = _validate.ValidateResource(goal, false);
 
-            Assert.IsInstanceOf<List<ValidationResult>>(result);
-            Assert.IsNotNull(result);
-            Assert.AreEqual(1, result.Count);
+            Assert.That(result, Is.InstanceOf<List<ValidationResult>>());
+            Assert.That(result, Is.Not.Null);
+            Assert.That(result.Count, Is.EqualTo(1));
         }
 
 
